feat: added withdrawal

+++ internal/models/accrual.go
@@ -1,7 +1,7 @@
 package models
 
 type Accrual struct {
-	OrderId string `json:"order"`
-	Status  string `json:"status"`
-	Accrual *int   `json:"accrual,omitempty"`
+	OrderId string   `json:"order"`
+	Status  string   `json:"status"`
+	Accrual *float64 `json:"accrual,omitempty"`
 }
